CL-0003
Rearranged Config files and Server file
