+++ config/env.js
@@ -1,5 +1,7 @@
-const env = {
-	APP_LINK : "/",
-	PASSWORD : "json"
+module.exports = {
+	USER_NAME : "root",
+	PASSWORD : "t",
+	PORT : 27017,
+	DATABASE : "begin",
+	secretOrKey: 'secret'
 }
-export default env;
