--- config/keys.js
@@ -1,4 +1,0 @@
-module.exports = {
-  mongoURI: 'mongodb://root:t@localhost:27017/begin',
-  secretOrKey: 'secret'
-};
