+++ config/passport.js
@@ -1,14 +1,19 @@
 const JwtStrategy = require('passport-jwt').Strategy;
 const ExtractJwt = require('passport-jwt').ExtractJwt;
+const passport = require('passport');
 const mongoose = require('mongoose');
-const User = mongoose.model('users');
-const keys = require('../config/keys');
+
+const User = require('../module/models/User')//mongoose.model('users');
+const keys = require('../config/env');
 
 const opts = {};
 opts.jwtFromRequest = ExtractJwt.fromAuthHeaderAsBearerToken();
 opts.secretOrKey = keys.secretOrKey;
 
-module.exports = passport => {
+module.exports = app => {
+  // Passport middleware
+  app.use(passport.initialize());
+
   passport.use(
     new JwtStrategy(opts, (jwt_payload, done) => {
       User.findById(jwt_payload.id)
