+++ server.js
@@ -1,50 +1,15 @@
 const express = require('express');
-const mongoose = require('mongoose');
-const bodyParser = require('body-parser');
-const passport = require('passport');
-const path = require('path');
-
-const index = require('./module/routes/index');
-const users = require('./module/routes/users');
-const contactList = require('./module/routes/contactList');
 
 const app = express();
-app.use('/public', express.static('public'));
 
-// Body parser middleware
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(bodyParser.json());
+require('./config/bodyParser')(app); // body Parser Config for Request Body
+require('./config/mongo'); // MongoDB DB Config
+require('./config/passport')(app); // Passport Config
 
-// DB Config
-const db = require('./config/keys').mongoURI;
-
-// Connect to MongoDB
-mongoose
-  .connect(db, { useNewUrlParser: true })
-  .then(() => console.log('MongoDB Connected'))
-  .catch(err => console.log(err));
-
-// Passport middleware
-app.use(passport.initialize());
-
-// Passport Config
-require('./config/passport')(passport);
-
-// Use Routes
-app.use('/api/users', users);
-app.use('/api/contact-list', contactList);
-app.use('/', index);
+require('./module/routes/index')(app, express); // Use Routes
 
 // Server static assets if in production
-if (process.env.NODE_ENV === 'production') {
-  // Set static folder
-  app.use(express.static('client/build'));
-
-  app.get('*', (req, res) => {
-    res.sendFile(path.resolve(__dirname, 'client', 'build', 'index.html'));
-  });
-}
+if (process.env.NODE_ENV === 'production') {}
 
 const port = process.env.PORT || 5000;
-
 app.listen(port, () => console.log(`Server running on port ${port}`));
